Add files via upload
I updated the loader to read file into disk. I also created a PCB class that acts as a node and a PCB_list class that acts as the linked list. While reading into disk, I added the address of the instruction into an array called jobInstructions and the data into an array called dataDisk for each new job. There is also a ID for every PCB that is the same as the jobID in the file. Next I will update the long term scheduler class that will load the  jobs and their data into the ram and also update the PCB as ram is loaded.

There is an issue with the loader where it sometimes gets a bad_alloc error. I'm still trying to figure this out but thats all I got for this weekend